Move files to better namespaces
