+++ Source code/MscModApi/MscModApi.csproj
@@ -81,28 +81,28 @@
     <Compile Include="Parts\BasicPart.cs" />
     <Compile Include="Parts\DerivablePart.cs" />
     <Compile Include="Parts\FsmPartData.cs" />
-    <Compile Include="Parts\PartEvent.cs" />
-    <Compile Include="Parts\PartEventListenerCollection.cs" />
+    <Compile Include="Parts\EventSystem\PartEvent.cs" />
+    <Compile Include="Parts\EventSystem\PartEventListenerCollection.cs" />
     <Compile Include="Parts\ReplacePart\GamePart.cs" />
     <Compile Include="Parts\ReplacePart\NewPart.cs" />
     <Compile Include="Parts\ReplacePart\OldPart.cs" />
-    <Compile Include="Parts\PartEventListener.cs" />
-    <Compile Include="Parts\ReplacePart\ReplacedPartEventListener.cs" />
-    <Compile Include="Parts\ReplacePart\ReplacedGamePartsEvent.cs" />
+    <Compile Include="Parts\EventSystem\PartEventListener.cs" />
+    <Compile Include="Parts\ReplacePart\EventSystem\ReplacedPartEventListener.cs" />
+    <Compile Include="Parts\ReplacePart\EventSystem\ReplacedGamePartsEvent.cs" />
     <Compile Include="Parts\ReplacePart\ReplacedGameParts.cs" />
     <Compile Include="Parts\ReplacePart\ReplacementPart.cs" />
-    <Compile Include="Parts\Box.cs" />
-    <Compile Include="Parts\BoxLogic.cs" />
-    <Compile Include="Parts\Kit.cs" />
+    <Compile Include="Parts\PartBox\Box.cs" />
+    <Compile Include="Parts\PartBox\BoxLogic.cs" />
+    <Compile Include="Parts\PartBox\Kit.cs" />
     <Compile Include="Parts\ReplacePart\SatsumaGamePart.cs" />
-    <Compile Include="Parts\SupportsPartBehaviourEvents.cs" />
-    <Compile Include="Parts\ReplacePart\SupportsReplacedGamePartsEvent.cs" />
-    <Compile Include="Parts\SupportsPartEvents.cs" />
+    <Compile Include="Parts\EventSystem\SupportsPartBehaviourEvents.cs" />
+    <Compile Include="Parts\ReplacePart\EventSystem\SupportsReplacedGamePartsEvent.cs" />
+    <Compile Include="Parts\EventSystem\SupportsPartEvents.cs" />
     <Compile Include="Saving\GamePartSave.cs" />
     <Compile Include="Saving\SerializableQuaternion.cs" />
     <Compile Include="Saving\SerializableVector3.cs" />
     <Compile Include="Shopping\ModItem.cs" />
-    <Compile Include="Parts\PartBox.cs" />
+    <Compile Include="Parts\PartBox\PartBox.cs" />
     <Compile Include="Shopping\ShopBaseInfo.cs" />
     <Compile Include="Shopping\Shop.cs" />
     <Compile Include="Shopping\ShopInterface.cs" />
@@ -113,7 +113,7 @@
     <Compile Include="Tools\Logger.cs" />
     <Compile Include="Tools\ScrewPlacementAssist.cs" />
     <Compile Include="Tool.cs" />
-    <Compile Include="Trigger\TriggerWrapper.cs" />
+    <Compile Include="Parts\Trigger\TriggerWrapper.cs" />
     <Compile Include="Tools\UserInteraction.cs" />
     <Compile Include="Tools\Helper.cs" />
     <Compile Include="MscModApi.cs" />
@@ -122,7 +122,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Parts\PartSave.cs" />
     <Compile Include="Parts\Screw.cs" />
-    <Compile Include="Trigger\Trigger.cs" />
+    <Compile Include="Parts\Trigger\Trigger.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\.editorconfig">
