+++ Source code/MscModApi/Parts/EventSystem/PartEvent.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MscModApi.Parts
+namespace MscModApi.Parts.EventSystem
 {
 	public static class PartEvent
 	{
