+++ Source code/MscModApi/Parts/ReplacePart/GamePart.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using HutongGames.PlayMaker;
 using MSCLoader;
 using MscModApi.Caching;
+using MscModApi.Parts.EventSystem;
 using MscModApi.Saving;
 using MscModApi.Tools;
 using UnityEngine;
