+++ Source code/MscModApi/Parts/ReplacePart/SatsumaGamePart.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using HutongGames.PlayMaker;
 using MscModApi.Caching;
+using MscModApi.Parts.EventSystem;
 using UnityEngine;
 
 namespace MscModApi.Parts.ReplacePart
