+++ Source code/MscModApi/Shopping/ShopItem.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using MscModApi.Parts;
+using MscModApi.Parts.PartBox;
 using MscModApi.Tools;
 using UnityEngine;
 using UnityEngine.UI;
