+++ Source code/MscModApi/Tools/Extensions.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using HutongGames.PlayMaker;
+using MscModApi.Parts.EventSystem;
 using MscModApi.Parts.ReplacePart;
 using UnityEngine;
